Snowball

Snowball.Stemmer

Snowball.release

Snowball.stem

Snowball.stem_all

Snowball.stemmer_types

Introduction

This package wraps the Snowball stemming library, originally developed by Dr Martin Porter. This package wraps the v2.0.0 version of the library.

Stemming algorithms

Currently, the following stemming algorithms are supported. The list of available algorithms can be found by using the stemmer_types() function in the package. By default, the UTF-8 version of the stemmers are used, but the ISO88591 versions are available in the underlying library if required.

arabic
basque
catalan
danish
dutch
english
finnish
french
german
greek
hindi
hungarian
indonesian
irish
italian
lithuanian
nepali
norwegian
porter
portuguese
romanian
russian
spanish
swedish
tamil
turkish

Snowball.Stemmer — Type

Stemmer(stemmer_type, charenc=UTF_8)

The Stemmer object stores a reference to the underlying libstemmer handle

Snowball.release — Method

release(stemmer::Stemmer)

Release resources held by libstemmer

Snowball.stem — Method

stem(stemmer::Stemmer, str)
stem(stemmer::Stemmer, words::Vector)

Stem the input with the Stemming algorthm of stemmer.

See also: stem_all

Snowball.stem_all — Method

stem_all(stemmer::Stemmer, sentence)

tokenize the string with the default tokenizer, and then stem each word

Snowball.stemmer_types — Method

stemmer_types()

List all the stemmer algorithms loaded.